Refactor : 세부 TestConfig 분리
