+++ src/test/java/shopping/config/TestConfig.java
@@ -1,14 +1,11 @@
 package shopping.config;
 
 import java.time.Duration;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.web.client.RestTemplate;
-import shopping.infrastructure.CurrencyLayerExchangeRateApi;
-import shopping.infrastructure.ExchangeRateApi;
 
 @Configuration
 @Profile("test")
@@ -16,14 +13,6 @@ public class TestConfig {
 
     private static final int CONNECTION_TIME_OUT = 5000;
     private static final int READ_TIME_OUT = 5000;
-    public static final int BACK_OFF_PERIOD = 1000;
-    private static final int MAX_ATTEMPT = 3;
-
-    @Value("${currency.accessKey}")
-    private String accessKey;
-    @Value("${currency.baseUrl}")
-    private String baseurl;
-
 
     @Bean
     public RestTemplateBuilder restTemplateBuilder() {
@@ -40,13 +29,4 @@ public class TestConfig {
             .build();
     }
 
-
-    @Bean
-    public ExchangeRateApi exchangeRateApi() {
-        return new CurrencyLayerExchangeRateApi(
-            accessKey,
-            baseurl,
-            restTemplate(restTemplateBuilder())
-        );
-    }
 }
